Revised all routes in api folder. Not a done job yet, need to fix the file paths.
